build: move signing key setup logic to a plugin
